Remove the #req helper method
I think this was a bad idea. It looks kinda weird and obscures the
intent. Consider:

req id: 4

vs:

subject.params = { id: 4 }
subject.call

The latter is clearly more verbose, but it's much more explicit about
what's being done, and makes it obvious that we are operating on the
object under test.

Closes #11